Add FTP support

+++ entrypoint.sh
@@ -7,18 +7,27 @@ USERNAME="${INPUT_USERNAME}"
 PASSWORD="${INPUT_PASSWORD:-""}"
 SSH_KEY="${INPUT_SSH_KEY:-""}"
 REMOTE_FOLDER="${INPUT_REMOTE_FOLDER:-"./"}"
-REPO_FOLDER="${INPUT_REPO_FOLDER:-"./"}"
+LOCAL_FOLDER="${INPUT_LOCAL_FOLDER:-"./"}"
 PORT="${INPUT_PORT:-"21"}"
-TYPE="${INPUT_TYPE:-"FTP"}"
+TYPE="${INPUT_TYPE:-"ftp"}"
 PUT_GITHUB_SHA="${INPUT_PUT_GITHUB_SHA:-"yes"}"
 
+if [ -z "${HOST}" ]; then
+    echo "Host is not defined"
+    exit -98
+fi
+
+if [ -z "${USERNAME}" ]; then
+    echo "Username is not defined"
+    exit -97
+fi
+
+cd "${LOCAL_FOLDER}"
 rm -rf .git
 rm -rf .github
 
-cd "${REPO_FOLDER}"
-
 if [ "${PUT_GITHUB_SHA}" == "yes" ]; then
-    touch ".github-putsftpaction-${GITHUB_SHA}"
+    echo "${GITHUB_SHA}" > ".github-put-sftp-action"
 fi
 
 echo ""
@@ -29,8 +38,8 @@ echo "--- List this directory ---"
 ls -lha
 echo ""
 
-BATCH_FILE=/tmp/commands
-if [ "${TYPE}" == "SFTP" ]; then
+if [ "${TYPE}" == "sftp" ]; then
+    BATCH_FILE=/tmp/commands
     echo "cd ${REMOTE_FOLDER}" > ${BATCH_FILE}
     echo "put -r ." >> ${BATCH_FILE}
     echo "ls" >> ${BATCH_FILE}
@@ -39,11 +48,10 @@ if [ "${TYPE}" == "SFTP" ]; then
     CMD=""
 
     if [ -n "${PASSWORD}" ]; then
-        echo "Len of password if ${#PASSWORD}"
-        CMD="sshpass -p \"${PASSWORD}\""
+        CMD="sshpass -p \"${PASSWORD}\" "
     fi
 
-    CMD="${CMD} sftp -P ${PORT} -oStrictHostKeyChecking=no -oBatchMode=no -b ${BATCH_FILE}"
+    CMD="${CMD} sftp -P ${PORT} -oStrictHostKeyChecking=no -oBatchMode=no -b ${BATCH_FILE} "
 
     if [ -n "${SSH_KEY}" ]; then
         export SSH_AUTH_SOCK=/tmp/ssh_agent.sock
@@ -52,7 +60,7 @@ if [ "${TYPE}" == "SFTP" ]; then
         ssh-add - <<< "${SSH_KEY}"
     fi
     
-    CMD="${CMD} ${USERNAME}@${HOST}"
+    CMD="${CMD} ${USERNAME}@${HOST} "
     
     echo "Command is:"
     echo "${CMD}"
@@ -60,8 +68,20 @@ if [ "${TYPE}" == "SFTP" ]; then
     ${CMD}
     
     EXIT_STATUS=$?
-elif [ "${TYPE}" == "FTP" ]; then
-    echo "Sviluppare FTP"
+elif [ "${TYPE}" == "ftp" ]; then
+    AUTH="${USERNAME}"
+    if [ -n "${PASSWORD}" ]; then
+        AUTH="${AUTH},${PASSWORD}"
+    fi
+
+    CMD="lftp --debug -e \"set ftp:ssl-allow no; set ftp:use-feat no; mirror -R ./ ${REMOTE_FOLDER}; quit;\" -u ${AUTH} -p ${PORT} ${HOST}"
+
+    echo "Command is:"
+    echo "${CMD}"
+
+    lftp --debug -e "set ftp:ssl-allow no; set ftp:use-feat no; mirror -R ./ ${REMOTE_FOLDER}; quit;" -u ${AUTH} -p ${PORT} ${HOST}
+    
+    EXIT_STATUS=$?
 fi
 
 echo "Exit status is: ${EXIT_STATUS}"
